Finish implementing swww support and fix cl argument parsing.

+++ src/common.rs
@@ -160,6 +160,6 @@ pub struct Cli {
     /// Restore previously set wallpapers
     pub restore: bool,
     #[arg(short, long, default_value_t = 0)]
-    /// How many error, warning, info, debug or trace logs will be shown. 0 for error, 1 for warning, 2 for info, 3 for debug, 4 or heigher for trace.
-    pub verbosity: usize,
+    /// How many error, warning, info, debug or trace logs will be shown. 0 for error, 1 for warning, 2 for info, 3 for debug, 4 or higher for trace.
+    pub verbosity: u8
 }
